1.5update

+++ AARS/AARS/Properties/FuelDrogue.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 public class FuelDrogue : PartModule
 {
     [KSPField]
-    public float FuelGain = 0f;
+    public double FuelGain = 0f;
     private Transform TirggerTransform = null;
     public bool IsDeployed = false;
     public bool IsContact = false;
@@ -76,7 +76,7 @@ public class FuelDrogue : PartModule
 
             if (IsContact && IsDeployed)
             {
-                float consumption = Probe.DrainSpeed * TimeWarp.fixedDeltaTime;
+                double consumption = Probe.DrainSpeed * TimeWarp.fixedDeltaTime;
                 FuelGain = this.part.RequestResource(Probe.fuelType, consumption);
                 if (consumption > FuelGain)
                     Probe.Deactivate();
